requirements.txt
alembic==1.5.8
amqp==5.0.6
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.7
billiard==3.6.4.0
bleach==3.3.0
blinker==1.4
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coverage==5.5
distlib==0.3.1
docopt==0.6.2
docutils==0.17.1
dominate==2.6.0
filelock==3.0.12
flake8==3.9.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-JWT-Extended==4.1.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==23.0.1
kombu==5.0.2
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
nose==1.3.7
packaging==20.9
pkginfo==1.7.0
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.15.0
SQLAlchemy==1.4.13
toml==0.10.2
tqdm==4.60.0
twine==3.4.1
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.6
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.4.1
